% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hadis_influence.R
\name{calculate_hadis_influence}
\alias{calculate_hadis_influence}
\title{Calculate Hadi's Influence Measure}
\usage{
calculate_hadis_influence(X, y)
}
\arguments{
\item{X}{A numeric matrix representing the design matrix (including the intercept term).}

\item{y}{A numeric vector representing the response variable.}
}
\value{
A numeric vector containing Hadi's Influence Measure for each observation.
}
\description{
This function calculates Hadi's Influence Measure for each observation in the
dataset, which is a measure of the influence of each observation on the
fitted regression model.
}
\keyword{internal}
