% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cooks_distance.R
\name{calculate_cooks_distance}
\alias{calculate_cooks_distance}
\title{Calculate Cook's Distance Measure}
\usage{
calculate_cooks_distance(X, y)
}
\arguments{
\item{X}{A numeric matrix representing the design matrix (including the intercept term).}

\item{y}{A numeric vector representing the response variable.}
}
\value{
A numeric vector containing Cook's Distance Measure for each observation.
}
\description{
This function calculates Cook's Distance Measure for each observation in the
dataset, which is a measure of the influence of each observation on the
fitted regression model.
}
\keyword{internal}
